Reader Social: gate the embed too behind content warnings (CM-626)
Previously the CW gate lived inside PostCardBody and only covered the
text body — image / video / audio / gifv attachments rendered
unconditionally. NSFW media on a sensitive post was fully visible.

Lift the gate up to SocialPostCard as a thin ContentWarningGate
wrapper that wraps both PostCardBody and PostCardEmbed. The header,
timestamp, and counts row still render so the user knows the post
exists and can read the spoiler text; clicking "Show content"
reveals body + embed together.

Class rename .social-post-card-body--gated -> .social-post-card-content-warning
because the gate is no longer body-specific.
