+++ client/reader/social/components/post-card/style.scss
@@ -10,7 +10,7 @@
 	.social-post-card-embed-images__item,
 	.social-post-card-embed-audio,
 	.social-post-card-embed-gifv,
-	.social-post-card-body--gated .components-button {
+	.social-post-card-content-warning .components-button {
 		position: relative;
 		z-index: 1;
 	}
